Déplacement de fonctions utilitaires dans un fichier helper
